demos/a2a_restaurant_rater/requirements.txt
a2a-sdk>=0.3.0
starlette>=0.37.0
uvicorn>=0.29.0
